fix: moved client state
