+++ libraries/rust/common/Cargo.toml
@@ -22,6 +22,8 @@ cosmwasm-schema = "1.2.1"
 sha2 = { version = "0.10.6", default-features = false }
 hex ={ version = "0.4.3", default-features = false }
 sha3 = { version = "0.10.6", default-features = false }
+serde-json-wasm = {version="0.5.0", default-features = false}
+ibc = { version = "0.32.0", default-features = false, features = ["parity-scale-codec", "mocks-no-std", "serde", "borsh"]}
 
 [dev-dependencies]
 test-utils={path="../test-utils"}
