+++ libraries/rust/common/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod btp_header;
+pub mod client_state;
 pub mod constants;
 pub mod icon;
 pub mod rlp;
